Remove complex TranslationResolverChain in favor of customizable base class
The TranslationResolverChain is removed to be more flexible with less
complexity when working with TranslationResolver. With the new implementation,
the Toolbox class receives a TranslationResolver with a common default setting
which reads from a messages.properties bundle. If more or other TranslationResolver
are needed, they can be added via a custom factory to be accessed directly by
their name.